Add dice rolls, fix error spam
This uses the API to address
https://github.com/rjb25/5e-dm-assistant/issues/7,
and also fixes a breaking function signature bug.

This also adds a way to detect error spam and
prompt the user for whether to continue in
the main interactive CLI loop.

Web App usage and CLI usage may begin to diverge
a bit more than before, although on the plus side
neither should break the other.